Add spambot honeypot and correct guests changing members
- Found the code responsible for guest editing member's comment, was an
implementation mistake. Switched inputs for divs, updated controller
functions. Fixed.
- Also updated robots.txt to disallow all bots and added a captcha
"alternative" - a hidden form field not shown to users ("honeypot") . If
it is filled with any input, then either a user was trying to edit it
manually to tamper with the page or a bot filled it out automatically -
in both cases we simply reject any change received. Will not cover all
cases, but should get rid of the simple spambots. References #92, but
need to implement in lara-2.0-beta later.
- Also fixed unknown bug: by wrong password the message for the user
said "changes saved" instead of "wrong password".