[Localization] Adds localization for erros
Added translations from typescript CreateEditEvent.ts to de.js and
en.js.

Also reworked how the errors were detected. Previously it was just
a long chain of if statements. Now we have an object which has the
translation keys as keys and values true or false (whether or not that
error should be displayed to the user). The keys are then passed through
the translate function to get the desired result.

If some new errors come up (to be expected with the integration of new
clubs) one has to only add a line containing the translation key and the
condition to the conditions object to update the errors.