[Statistics] Adds caching of preferred statistics
If a user selects a club to show the statistics we now
save this preference to localStorage. On page load, query
the storage if such a preference has been saved, and if so
select the corresponding statistics by clicking the correct link.

It would probably be cleaner if the server stored said preferences
and would set the corresponding statistics as active, as clicking
the link manually will result in a visible switch of the presented
table. One could make an ajax request to a suitable Controller
and use the saved value (on the server) in statisticsMembers.blade.php
to set the correct list item to active.

Will also add this functionality to the leaderboards.