[Statistics] Save sorting preferences
We now write the preferred sorting (by name or by amounts of shifts?)
and the order (ascending or descending?) to the local storage.

From there they are retrieved on page load.

To apply the sorting on page load, the sorting and updating of the rows
and icons was moved to separate functions.

Also the stored sort order is applied to all tables in the member
statistics view (so also hidden ones). It would probably be more
consistent to add this behaviour to the normal sorting, too.