[Statistics] Show shifts in other clubs
This commit adds the shifts that a user has done in a club that is not
his own to the monthly statistic. The other shifts (if present) are
displayed in grey behind the users shifts in his own club. They are also
shown in the progress bar in grey.

Sorting of shifts should still work, as JS will just parse the first
number (stopping at the first space). In the future one may want to add
an option to sort by total amount of shifts. Then it will become
necessary to calculate the total amount of shifts done in a club on the
frontend. One could try to eval the text in the shifts column to get the
total amount of shifts.

A tooltip indicating what the hidden shifts are should also be added so
users know what the grey numbers mean.