[Statistics] Change leaderboards to whole semester
Previously leaderboards showed only the top 10 shifts for the current
month. Now we fetch the data for the whole semester.

I am creating a new feature branch because this makes loading times
considerably longer for statistics view (as expected), making it
excruciating to switch between different months.

This should probably be solved as an ajax call, which then saves the
current leaderboard to local storage, displaying it again on page load.
This should easily fit in local storage, as there are 20 columns at
maximum. We can then reload the leaderboard after the ajax has finished.

Maybe there is an easier solution which requires less computing power
server side, but as of right now I don't see how. This would be the
preferred way to solve this.