[Events] Made section dropdown readonly
Adding the 'readonly' property to the laravel form to disable text input
from users.

In the future it would probably be better to use a real dropdown (see
[Bootstrap Components](http://getbootstrap.com/components/#dropdowns)).
I tried that but ran into problems with submitting the correct place
value to the server. For now this is the quicker but maybe not so clean
solution.