[Statistics] Make date format more readable
Using strfttime we display the date in a more readable way. It is
debatable whether '%a, %d.%m' would result in a better format (%a
standing for the weekday). The problem with this format is, that the
server language somehow is stuck to english and strfttime will always
output the enlish names for the weekday (Monday and Tuesday, not Montag
and Dienstag). Until we fix this (also necessary for event and month view)
I would not display the day of the week.