[Event View] Fix status icon update
Previously, when selecting "I'll do it" from the dropdown, the status
icon of the user would not update. This was due to a faulty calculation
of the currentId of the shift in the beforeSend() of the ajax call. The
id would be calculated via the submit button, which was not always
present. When not present, the id would be set to undefined, causing the
spinner to not appear, which then could not be replaced bt an icon.
Extracting the id from the button would only work when there was
actually a button, which is not the case if nothing was typed in the
input for the shift.

To fix this, we now extract the id of the shift before the ajax call
starts, so that it is always available when necessary.