[IDE_Helper] Remove calls to non-existant class
The app/Revision_SurveyAnswwer.php class was deleted in c5ee950, but
some function calls still used it. This resulted in a failure in
php artisan ide_helper:models, which tried to acces this class.

To generate the models file we need another package (see https://github.com/barryvdh/laravel-ide-helper#automatic-phpdocs-for-models), which we now include.

To generate models with homestead use
```
vagrant ssh
composer update
php artisan ide-helper:models
(exit)
```
(Model information is generated using the DB, thus we need acces to the
DB which we only have on the VM).