New policy for creating ShiftTypes
When creating a new shift we search the database for a shifttype that
has the same title, start and end time as the new shift. If none exits.
we create a new shifttype. If one exists, no new shifttype will be
created.

If the time/end of this shift is changed afterwards, we will not create
a new shifttype (as long as the title of the shift does not change).