Fix event view butons not working
Previously the buttons were only actived for week view. This is fixed now
Nevertheless we should think about making a more abstract concept for
buttons that toggle things. This seems to only duplicate code right now
as we have to write the toggling for the week view and the event view
separately. Also this behaviour will likely be necesarry for other stuff
in the future.