Update method to determine active clubs
For statistics we evaluate the acitve clubs by looking at all sections
and returning the clubs that match those name.

In production DB having two clubs with the same name is not prohibited
(and oddly enough, there are two entities of 'bd-club' in the productio
DB).
This commit provides a quick-and-dirty solution to this problem: use the
club provided by the section (as this is the one that is used for the
user / person relationship).

In the future, we should refactor the StatisticsController to only use
Sections / Users (instead of clubs / persons ).