Add log message for synced bd-Events
We will now log the timestamp newly synced bd-Events.
Because this command will be run by the server, no user is logged in. To
prevent `Auth::user()` failure in the logging, we set the logged name to
'Lara'.

Please discuss if this is appropriate, other names could be 'System' or
'BD-Sync'.