Fix logging in events (details + shifts)
Since logging was implemented something broke the correct triggers for
logs. This PR aims to fix this.

First of, if one calls `$event->save()` before the `$event->isDirty(...)`
checks are called, the event is saved to the DB, therefore no attribute
is dirty. This caused the missing logs for event infos.

Line 60 in the ShiftController caused shifts to always be logged as
newly created. This is now fixed (see ~line 70 in the same file).