Fix/user created from person (#272)
* Fix guest privacy policy

Add a check for logged in user, so it does not throw null for guests

* Fix logging in events (details + shifts)

Since logging was implemented something broke the correct triggers for
logs. This PR aims to fix this.

First of, if one calls `$event->save()` before the `$event->isDirty(...)`
checks are called, the event is saved to the DB, therefore no attribute
is dirty. This caused the missing logs for event infos.

Line 60 in the ShiftController caused shifts to always be logged as
newly created. This is now fixed (see ~line 70 in the same file).

* Add log message for synced bd-Events

We will now log the timestamp newly synced bd-Events.
Because this command will be run by the server, no user is logged in. To
prevent `Auth::user()` failure in the logging, we set the logged name to
'Lara'.

Please discuss if this is appropriate, other names could be 'System' or
'BD-Sync'.

* log messages are caused by a select that gives duplicated results, make it unique will fix the problem

* Fix user status not transfered to person

* Fix person status in update method too

* createFromPerson now adds to $person

* Fix user created from event view

Was caused by Person::fullName creating a user

* Fix statistics user access

* Adapt migrations to new user() call