Fix infinite loop
Before this commit, an infinite loop could occur: when logging changes
in the schedule, we also save the corresponding schedule, thus creating
an infinite loop. This is circumvented here by not saving the schedule
by default in the logging function.

This works, but it is quite an ugly fix. A better solution would be to
extract the whole logging / revisions shenanigangs into a separate
table, which we could then use to log changes to events / surveys /
schedules etc.. One could then use the Laravel polymorphic relationships
to morph between those. A relict from an old SWP is found in the
revisions table, maybe one could adapt this solution?